+++ readers/binary.js
@@ -1,18 +1,14 @@
 'use strict'
 
-const fs = require('fs')
-
 exports.load = (name) => {
-  return fs.readFileSync(name)
+  return require('node:fs').readFileSync(name)
 }
 
-exports.loadPromise = (name) => {
-  return new Promise((resolve, reject) => {
-    fs.readFile(name, (err, content) => {
-      if (err) return reject(err)
-      resolve({ path: name, data: content })
-    })
-  })
+exports.loadPromise = async (name) => {
+  return {
+    path: name,
+    data: await require('node:fs/promises').readFile(name),
+  }
 }
 
 exports.empty = () => {