+++ readers/js.js
@@ -4,6 +4,10 @@ exports.load = (name) => {
   return require(name)
 }
 
+exports.loadPromise = async (name) => {
+  return require(name)
+}
+
 exports.empty = () => {
   return {}
 }