+++ readers/yaml.js
@@ -1,10 +1,13 @@
 'use strict'
 
-const fs = require('fs')
 const yaml = require('js-yaml')
 
 exports.load = (name) => {
-  return yaml.load(fs.readFileSync(name, 'utf8'))
+  return yaml.load(require('node:fs').readFileSync(name, 'UTF-8'))
+}
+
+exports.loadPromise = async (name) => {
+  return yaml.load(await require('node:fs/promises').readFile(name, 'UTF-8'))
 }
 
 exports.empty = () => {